Move some funcs from api to rag module
